Move mutation and query to types.py
